+++ src/plugins/contentRenderer/table/header/basic/tableHeaderContentRenderer.component.ts
@@ -4,6 +4,8 @@ import {extend} from "@jscrpt/common";
 import {TableHeaderContentRendererOptions} from "../../tableContentRenderer.interface";
 import {HEADER_CONTENT_RENDERER_OPTIONS} from "../../../types";
 import {HeaderContentRendererAbstractComponent} from "../../../headerContentRendererAbstract.component";
+import {GRID_PLUGIN_INSTANCES} from '../../../../../components/grid/types';
+import {GridPluginInstances} from '../../../../../components/grid';
 
 /**
  * Default options for 'TableHeaderContentRendererComponent'
@@ -34,25 +36,8 @@ const defaultOptions: TableHeaderContentRendererOptions =
 {
     selector: 'thead.content-renderer',
     templateUrl: 'tableHeaderContentRenderer.component.html',
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    styles:
-    [
-        `.header-default
-        {
-            white-space: nowrap;
-        }
-
-        .header-orderable:hover
-        {
-            background-color: #E3E3E3;
-            cursor: pointer;
-        }
-
-        .header-content
-        {
-            white-space: normal;
-        }`
-    ]
+    styleUrls: ['tableHeaderContentRenderer.component.css'],
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class TableHeaderContentRendererComponent<TData> extends HeaderContentRendererAbstractComponent<TData, TableHeaderContentRendererOptions>
 {
@@ -70,9 +55,10 @@ export class TableHeaderContentRendererComponent<TData> extends HeaderContentRen
     //######################### constructor #########################
     constructor(pluginElement: ElementRef,
                 changeDetector: ChangeDetectorRef,
-                @Inject(HEADER_CONTENT_RENDERER_OPTIONS) @Optional() options: TableHeaderContentRendererOptions)
+                @Inject(HEADER_CONTENT_RENDERER_OPTIONS) @Optional() options: TableHeaderContentRendererOptions,
+                @Inject(GRID_PLUGIN_INSTANCES) @Optional() gridPlugins: GridPluginInstances)
     {
-        super(pluginElement, changeDetector);
+        super(pluginElement, gridPlugins, changeDetector);
 
         this._options = extend(true, {}, defaultOptions, options);
     }